ready

+++ src/main/java/com/mieshkov/corplan/model/dao/GenericDao.java
@@ -9,13 +9,13 @@ public interface GenericDao<T> extends AutoCloseable {
 
     void create(T entity);
 
-    T findById(int id);
+    T findById(Long id);
 
     List<T> findAll();
 
     void update(T entity);
 
-    void delete(int id);
+    void delete(Long id);
 
     void close();
 }
